+++ MapboxDirections/MBCoordinateBounds.swift
@@ -6,16 +6,26 @@ import CoreLocation
  */
 @objc(MBCoordinateBounds)
 public class CoordinateBounds: NSObject, Codable {
-    let northWest: CLLocationCoordinate2D
-    let southEast: CLLocationCoordinate2D
+    let southWest: CLLocationCoordinate2D
+    let northEast: CLLocationCoordinate2D
+    
+    /**
+     Initializes a `BoundingBox` with known bounds.
+     */
+    @objc
+    public init(southWest: CLLocationCoordinate2D, northEast: CLLocationCoordinate2D) {
+        self.southWest = southWest
+        self.northEast = northEast
+        super.init()
+    }
     
     /**
      Initializes a `BoundingBox` with known bounds.
      */
     @objc
     public init(northWest: CLLocationCoordinate2D, southEast: CLLocationCoordinate2D) {
-        self.northWest = northWest
-        self.southEast = southEast
+        self.southWest = CLLocationCoordinate2D(latitude: southEast.latitude, longitude: northWest.longitude)
+        self.northEast = CLLocationCoordinate2D(latitude: northWest.latitude, longitude: southEast.longitude)
         super.init()
     }
     
@@ -32,20 +42,20 @@ public class CoordinateBounds: NSObject, Codable {
         var minimumLongitude: CLLocationDegrees = 180
         
         for coordinate in coordinates {
-            maximumLatitude = Swift.max(maximumLatitude, coordinate.latitude)
-            minimumLatitude = Swift.min(minimumLatitude, coordinate.latitude)
-            maximumLongitude = Swift.max(maximumLongitude, coordinate.longitude)
-            minimumLongitude = Swift.min(minimumLongitude, coordinate.longitude)
+            maximumLatitude = max(maximumLatitude, coordinate.latitude)
+            minimumLatitude = min(minimumLatitude, coordinate.latitude)
+            maximumLongitude = max(maximumLongitude, coordinate.longitude)
+            minimumLongitude = min(minimumLongitude, coordinate.longitude)
         }
         
-        let northWest = CLLocationCoordinate2D(latitude: minimumLatitude, longitude: minimumLongitude)
-        let southEast = CLLocationCoordinate2D(latitude: maximumLatitude, longitude: maximumLongitude)
+        let southWest = CLLocationCoordinate2D(latitude: minimumLatitude, longitude: minimumLongitude)
+        let northEast = CLLocationCoordinate2D(latitude: maximumLatitude, longitude: maximumLongitude)
         
-        self.init(northWest: northWest, southEast: southEast)
+        self.init(southWest: southWest, northEast: northEast)
     }
     
-    var path: String {
-        return "\(northWest.longitude),\(northWest.latitude);\(southEast.longitude),\(southEast.latitude)"
+    public override var description: String {
+        return "\(southWest.longitude),\(southWest.latitude);\(northEast.longitude),\(northEast.latitude)"
     }
 }